Refactoring; Moved deploy manager to sub, using difi-deploy-manager as parent pom
