+++ deploy-manager/pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>no.difi.deploymanager</artifactId>
+    <groupId>difi-deploy-manager</groupId>
+    <version>0.9.1-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+
+  <groupId>no.difi.deploymanager</groupId>
+  <artifactId>deploy-manager</artifactId>
+  <version>0.9.1-SNAPSHOT</version>
+
+</project>
