+++ deploy-manager/src/main/java/no/difi/deploymanager/artifact/Application.java
