+++ deploy-manager/src/main/java/no/difi/deploymanager/domain/ApplicationData.java
