+++ deploy-manager/src/main/java/no/difi/deploymanager/domain/DownloadedVersion.java
