+++ deploy-manager/src/main/java/no/difi/deploymanager/domain/Status.java
