+++ deploy-manager/src/main/java/no/difi/deploymanager/download/dto/DownloadDto.java
