+++ deploy-manager/src/main/java/no/difi/deploymanager/remotelist/exception/RemoteApplicationListException.java
