+++ deploy-manager/src/main/java/no/difi/deploymanager/util/JsonUtil.java
