+++ deploy-manager/src/main/java/no/difi/deploymanager/versioncheck/dto/CheckVersionDto.java
