+++ deploy-manager/src/main/java/no/difi/deploymanager/versioncheck/exception/ConnectionFailedException.java
