+++ deploy-manager/src/test/java/no/difi/deploymanager/testutils/ObjectMotherApplicationData.java
