--- src/test/java/no/difi/deploymanager/domain/ApplicationDataTest.java
@@ -1,16 +1,0 @@
-package no.difi.deploymanager.domain;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertNotNull;
-
-public class ApplicationDataTest {
-    private ApplicationData data;
-
-    @Test
-    public void should_get_empty_download_list_when_not_initialized() {
-        data = new ApplicationData();
-
-        assertNotNull(data.getDownloadedVersions());
-    }
-}
